fix confusion when inspecting unbound slots
In the current state, INSPECT displays unbound slots as the "unbound" string.
This is confusing and error inducing. The problem can be visualized as
follows:

  (defclass foo () ((a :initarg :a) (b :initarg :b)))
  (make-instance 'foo :a "unbound")
  (inspect *)

  The object is a STANDARD-OBJECT of type FOO.
  0. A: "unbound"
  1. B: "unbound"

With this patch, INSPECT uses the same representation as DESCRIBE for unbound
slots, i.e. the representation of an unreadable object with label "unbound
slot".

With the previous example:

  The object is a STANDARD-OBJECT of type FOO.
  0. A: "unbound"
  1. B: #<unbound slot>

This removes all confusion and makes INSPECT consistent with DESCRIBE.